requirements.txt
flask
google-cloud-storage
google-cloud-bigquery
google-cloud-texttospeech
google-auth
python-dotenv
vertexai
